# feat: Hybrid Storage Backend - SQLite-vec Speed + Cloudflare Persistence